openpilot v0.9.9 release
date: 2025-04-11T09:10:03
master commit: 24c2d9a68a03f3282519c7ffd82241e38df305f5

+++ tinygrad_repo/tinygrad/nn/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 import math
-from tinygrad.tensor import Tensor, dtypes
+from tinygrad.tensor import Tensor
+from tinygrad.dtype import dtypes
 from tinygrad.device import is_dtype_supported
 from tinygrad.helpers import prod, make_tuple, flatten
 from tinygrad.nn import optim, state, datasets  # noqa: F401
@@ -322,7 +323,7 @@ class Embedding:
     if not hasattr(self, 'arange'): self.arange = Tensor.arange(self.vocab_sz, requires_grad=False, device=self.weight.device).unsqueeze(-1)
     big_shp = idx.shape+(self.vocab_sz, self.embed_sz)
     arange, idx, vals = self.arange.expand(big_shp), idx.reshape(idx.shape+(1, 1)).expand(big_shp), self.weight.expand(big_shp)
-    return (arange == idx).mul(vals).sum(-2, acc_dtype=vals.dtype)
+    return (arange == idx).mul(vals).sum(-2, dtype=vals.dtype)
 
 class LSTMCell:
   """
